refactoring and added navigation to program item detail view
