+++ IronRoadForChildren/Features/Program/Day/ProgramItemDelailView.swift
@@ -1,0 +1,20 @@
+//
+//  ProgramItemDelailView.swift
+//  IronRoadForChildren
+//
+//  Created by Robert Zavaczki on 18.04.23.
+//
+
+import SwiftUI
+
+struct ProgramItemDelailView: View {
+	var body: some View {
+		Text("Program Item Detail View")
+	}
+}
+
+struct ProgramItemDelailView_Previews: PreviewProvider {
+	static var previews: some View {
+		ProgramItemDelailView()
+	}
+}
